+++ scripts/keyboard.js
@@ -1,0 +1,18 @@
+document.addEventListener("keydown", (event) => {
+    const keyName = event.key;
+
+    if (keyName === "r") {
+        
+    }
+
+  },
+  false
+);
+
+document.addEventListener(
+  "keyup",
+  (event) => {
+    const keyName = event.key;
+  },
+  false
+);